backend/pyproject.toml
[tool.poetry]
name = "ck-missive-backend"
version = "0.1.0"
description = "Backend for CK_Missive project"
authors = ["bluefishs <125471205+bluefishs@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.34"
alembic = "^1.13.1"
pydantic = "^2.8.2"
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
python-dotenv = "^1.0.0"
celery = "^5.3.4"
redis = "^5.0.1"
notion-client = "^2.2.1"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
python-docx = "^1.1.0"
tqdm = "^4.66.1"
tenacity = "^8.2.3"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
psycopg2-binary = "^2.9.9" # Use binary for environments without build tools
# asyncpg = "^0.29.0" # Temporarily disabled due to build issues
httpx = "^0.27.0"
google-auth = "^2.29.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.181.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
